Changed up some of the UI elements to make them look a little better. On top of that, oh yes, WE HAVE SIGN IN. The sign works with the Google account set up on the phone with user. The best part, if they want to use a different log in, they can if it is registered with Google which is pretty awesome. If needed, we can include other sign in options such as Facebook, Twitter, and others if others so desire.
The app icon is also updated along with the title of the app itself. The app when you login will store user's log in information to the cache of the android device. If it does not, the user will have to sign in again which is super important. A token is used to create a login which is really secure thanks to Google.

When a user signs in, they are welcomed using their name and sent to the main game ui, which for right now just says "main game ui" so nothing to crazy yet, but hey, it's a wonderful start.