Adds very rough draft of database interactions DO NOT MERGE TO MASTER.
This is just to give you all some ideas on how to use the database. Look at the last lines of onCreate to see how to attach listeners and the top of openNotificationsFragment to see how to add to a list. The stall line in openCurrentGames is to inspect and verify the data update via listeners. Also the DBIOCore has been updated to closer resemble it's final form.