Massive UI Overhaul with problems fixed thanks to Miles
I have redone the UI completely for our presentation to match material design. I have also included nicer directories for managing the UI and have also added a directory for us to add working apk files to.

Features included:
- Material design
- Board with castles (colors not changed sorry guys)
- Added names and emails to drawer layout to let the user's know that they are signed in.
- Added a semi-working notification counter that doesn't work on startup until after a user has seen there available notifications.

Things that are broken:
- When the app is killed (by Android or by the user) we lose the variable currentUser. I don't know why exactly.
- Notification counter does not grab notifications that are newly sent on app startup until the notification button is pressed and then the numbers update. The best bet is to pull from the database in order to grab these and updated the counter during startup.

Please check it out and accept it when ready!